+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## 0.1.1
+- More documentation.
+- `failureStream` and `retryStream` helpers.
+- Fixed a bug with cleaning up failed tasks.
+
 ## 0.1.0
 - Initial release.